memory, zram: introduce and export readswap()
try_zram_decompress() is currently exported to extension modules, but
from a viewpoint of author of extension modules, it's better to export
an interface to read memory on swap; difference of decompressor are
then hidden within the interface and there is no need for extension
modules to update accordingly each time new decompressor are added in
the future.

So let's introduce function readswap() as an interface to read memory
on swap.

Signed-off-by: HATAYAMA Daisuke <d.hatayama@fujitsu.com>